+++ internal/basicchain/testdata/invoke/invokescript_contract.go
