+++ internal/basicchain/testdata/test_contract.go
