+++ pkg/core/native_management_test.go
@@ -3,6 +3,7 @@ package core_test
 import (
 	"testing"
 
+	"github.com/nspcc-dev/neo-go/internal/basicchain"
 	"github.com/nspcc-dev/neo-go/pkg/config"
 	"github.com/nspcc-dev/neo-go/pkg/core/native/nativenames"
 	"github.com/nspcc-dev/neo-go/pkg/neotest"
@@ -22,10 +23,10 @@ func TestManagement_GetNEP17Contracts(t *testing.T) {
 
 	t.Run("basic chain", func(t *testing.T) {
 		bc, validators, committee := chain.NewMultiWithCustomConfig(t, func(c *config.ProtocolConfiguration) {
-			c.P2PSigExtensions = true // `initBasicChain` requires Notary enabled
+			c.P2PSigExtensions = true // `basicchain.Init` requires Notary enabled
 		})
 		e := neotest.NewExecutor(t, bc, validators, committee)
-		initBasicChain(t, e)
+		basicchain.Init(t, "../../", e)
 
 		require.ElementsMatch(t, []util.Uint160{e.NativeHash(t, nativenames.Neo),
 			e.NativeHash(t, nativenames.Gas), e.ContractHash(t, 1)}, bc.GetNEP17Contracts())
